[#4423] Seperate DefaultApplicationContext
Separates the logic of the DefaultApplicationContext.
