+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/module/ModuleInstanceManager.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2017 NAVER Corp.
+ * Copyright 2018 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,11 +16,15 @@
 
 package com.navercorp.pinpoint.profiler.context.module;
 
-import com.navercorp.pinpoint.bootstrap.AgentOption;
-
 /**
- * @author Woonduk Kang(emeroad)
+ * @author Taejin Koo
  */
-public interface ModuleFactory {
-    PinpointModuleHolder newModule(AgentOption agentOption);
+interface ModuleInstanceManager {
+
+    void start();
+
+    ModuleInstanceHolder getModuleInstanceHolder();
+
+    void stop();
+
 }
